# slog: faster and less memory-consumption